+++ config/bref.php
@@ -1,0 +1,21 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Servable Assets
+    |--------------------------------------------------------------------------
+    |
+    | Here you can configure list of public assets that should be servable
+    | from your application's domain instead of only being servable via
+    | the public S3 "asset" bucket or the AWS CloudFront CDN network.
+    |
+    */
+
+    'assets' => [
+        // 'favicon.ico',
+        // 'robots.txt',
+    ],
+
+];