+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed 
+- Made `ServeStaticAssets` configurable [19fb1ac](19fb1ac21fd7245a8bd529eb6325cea2308ffbf2)
 
 ## [v0.1.0] - 2022-05-18
 ### Added